requirements.txt
tensorflow
google-generativeai
pillow
numpy
streamlit
#it works on python==3.12
#pip install tf-keras